+++ backend/hqlib/metric_source/checkmarx.py
@@ -50,14 +50,10 @@ class Checkmarx(MetricSourceWithIssues):
             super().__init__(name.replace('_', ' '))
 
     def __init__(self, url: str, username: str, password: str, *args, **kwargs) -> None:
-        self._sast_report_id = 0
-        self._last_scan_url = urllib.parse.urljoin(
-            url, '/Cxwebinterface/odata/v1/Projects?$expand=LastScan&$filter=Name%20eq%20%27{project_name}%27')
-        self._token_url = urllib.parse.urljoin(url, '/cxrestapi/auth/identity/connect/token')
-        self._token_post_body = \
-            'username={username}&password={password}&scope=sast_rest_api&grant_type=password' \
-            '&client_id=resource_owner_client&client_secret=014DF517-39D1-4453-B7B3-9930C563627C'\
-            .format(username=username, password=password)
+        self._projects_url = urllib.parse.urljoin(url, 'CxRestAPI/projects')
+        self._scan_url = urllib.parse.urljoin(url, '/CxRestAPI/sast/scans?projectId={project_id}&last=1')
+        self._statistics_url = urllib.parse.urljoin(url, '/CxRestAPI/sast/scans/{id}/resultsStatistics')
+
         self._generate_report_post_url = urllib.parse.urljoin(url, '/CxRestAPI/reports/sastScan')
         self._sast_scan_status_url = urllib.parse.urljoin(url, '/CxRestAPI/reports/sastScan/{scan_id}/status')
         self._sast_scan_url = urllib.parse.urljoin(url, '/CxRestAPI/reports/sastScan/{scan_id}')
@@ -65,8 +61,14 @@ class Checkmarx(MetricSourceWithIssues):
             url, '/CxWebClient/ScanQueryDescription.aspx?'
             'queryID={query_id}&queryVersionCode={query_version_code}&queryTitle={name}')
         self.__avoid_crl_check()
-        self._url_open = url_opener.UrlOpener(username=username, password=password)
-        self._token_authorised_url_open = url_opener.UrlOpener(authorization_token=self._retrieve_access_token())
+
+        token_url = urllib.parse.urljoin(url, '/cxrestapi/auth/identity/connect/token')
+        token_post_body = \
+            'username={username}&password={password}&scope=sast_rest_api&grant_type=password' \
+            '&client_id=resource_owner_client&client_secret=014DF517-39D1-4453-B7B3-9930C563627C' \
+            .format(username=username, password=password)
+        self._url_open = url_opener.UrlOpener(
+            authorization_token=self.__retrieve_access_token(token_url, token_post_body))
         super().__init__(url=url, *args, **kwargs)
 
     @staticmethod
@@ -82,11 +84,12 @@ class Checkmarx(MetricSourceWithIssues):
             # Handle target environment that doesn't support HTTPS verification
             ssl._create_default_https_context = _create_unverified_https_context
 
+    @staticmethod
     @functools.lru_cache(maxsize=4096)
-    def _retrieve_access_token(self) -> str:
+    def __retrieve_access_token(token_url: str, token_post_body: str) -> str:
         try:
             opener = url_opener.UrlOpener()
-            json_string = opener.url_read(self._token_url, post_body=bytes(self._token_post_body, 'ascii'))
+            json_string = opener.url_read(token_url, post_body=bytes(token_post_body, 'ascii'))
             return json.loads(json_string)['access_token']
         except url_opener.UrlOpener.url_open_exceptions:
             logging.error("HTTP error during the retrieving of access token!")
@@ -95,20 +98,21 @@ class Checkmarx(MetricSourceWithIssues):
         return ''
 
     def metric_source_urls(self, *report_urls: str) -> List[str]:
+        """ Return the url(s) to the metric source for the metric source id. """
         checkmarx_report_urls = []
-
         for project_name in report_urls:
             try:
-                last_scan_json = self._fetch_last_scan(project_name)
+                project_id = self._fetch_project_id(project_name)
+                last_scan_id = self._fetch_last_scan(project_id)[0]["id"]
                 checkmarx_report_urls.append(urllib.parse.urljoin(
                     self.url(),
                     "/CxWebClient/ViewerMain.aspx?scanId={}&ProjectID={}"
-                    .format(str(last_scan_json["Id"]), str(last_scan_json["ProjectId"]))))
-            except (IndexError, KeyError, TypeError) as reason:
+                    .format(str(last_scan_id), str(project_id))))
+            except (IndexError, KeyError) as reason:
                 logging.error("Couldn't load values from json: %s - %s", project_name, reason)
+                return [self.url()]
             except url_opener.UrlOpener.url_open_exceptions:
                 return [self.url()]
-
         return checkmarx_report_urls
 
     def nr_warnings(self, metric_source_ids: Iterable[str], priority: str) -> int:
@@ -116,9 +120,14 @@ class Checkmarx(MetricSourceWithIssues):
         nr_alerts = 0
         for project_name in metric_source_ids:
             try:
-                last_scan_json = self._fetch_last_scan(project_name)
-                nr_alerts += last_scan_json[priority.title()]
+                project_id = self._fetch_project_id(project_name)
+                last_scan_id = self._fetch_last_scan(project_id)[0]["id"]
+                statistics_json = self._fetch_statistics(last_scan_id)
+                nr_alerts += statistics_json[priority.lower() + 'Severity']
             except url_opener.UrlOpener.url_open_exceptions:
+                return -1
+            except StopIteration:
+                logging.error("Error: no project id found for project with name '%s'.", project_name)
                 return -1
             except (KeyError, IndexError) as reason:
                 logging.error("Couldn't parse alerts for project %s with %s risk level from %s: %s",
@@ -128,7 +137,7 @@ class Checkmarx(MetricSourceWithIssues):
 
     def _create_report(self, last_scan_id: int) -> int:
         post_request_body = 'reportType=XML&scanId={scan_id}'.format(scan_id=last_scan_id)
-        json_string = self._token_authorised_url_open.\
+        json_string = self._url_open.\
             url_read(self._generate_report_post_url, post_body=bytes(post_request_body, 'ascii'))
         return json.loads(json_string)['reportId']
 
@@ -136,7 +145,7 @@ class Checkmarx(MetricSourceWithIssues):
         """ Exception thrown when SAST report is not found in status ready after defined retries. """
 
     def _get_issues_from_report(self, sast_report_id: int, priority: str) -> List:
-        sast_report = self._token_authorised_url_open.url_read(self._sast_scan_url.format(scan_id=sast_report_id))
+        sast_report = self._url_open.url_read(self._sast_scan_url.format(scan_id=sast_report_id))
         root = xml.etree.cElementTree.fromstring(sast_report)
         issues = []
         queries = root.findall(".//Query")
@@ -169,14 +178,16 @@ class Checkmarx(MetricSourceWithIssues):
         self._issues = []
         for project_name in metric_source_ids:
             try:
-                last_scan_id = self._fetch_last_scan(project_name)['Id']
+                project_id = self._fetch_project_id(project_name)
+                last_scan_id = self._fetch_last_scan(project_id)[0]["id"]
                 logging.info("Retrieved last scan id: %s.", last_scan_id)
+                sast_report_id = self._create_report(last_scan_id)
+                logging.info("Retrieved SAST report id %s for scan with id %s.", sast_report_id, last_scan_id)
 
-                self._sast_report_id = self._create_report(last_scan_id)
-                logging.info("Retrieved SAST report id %s for scan with id %s.", self._sast_report_id, last_scan_id)
-
-                self._wait_for_sast_report()
-                self._issues = self._get_issues_from_report(self._sast_report_id, priority)
+                self._wait_for_sast_report(sast_report_id)
+                logging.info(
+                    "SAST report with id %s for scan with id %s created successfully.", sast_report_id, last_scan_id)
+                self._issues = self._get_issues_from_report(sast_report_id, priority)
             except url_opener.UrlOpener.url_open_exceptions:
                 pass
             except ValueError as reason:
@@ -188,17 +199,18 @@ class Checkmarx(MetricSourceWithIssues):
             except self.SastReportNotCreated:
                 logging.error('SAST report is not created on the Checkmarx server!')
 
-    def _wait_for_sast_report(self, repeat_nr: int = 5, wait_seconds: int = 3):
-        while self.__get_json(self._sast_scan_status_url.format(scan_id=self._sast_report_id),
-                              self._token_authorised_url_open)["status"]["value"] != "Created":
-            self._token_authorised_url_open.url_read.cache_clear()
+    def _wait_for_sast_report(self, sast_report_id: int, repeat_nr: int = 10, wait_seconds: int = 5):
+        time.sleep(2)
+        status_url = self._sast_scan_status_url.format(scan_id=sast_report_id)
+        while self.__get_json(status_url)["status"]["value"] != "Created":
+            self._url_open.url_read.cache_clear()
             repeat_nr -= 1
             if not repeat_nr:
                 raise self.SastReportNotCreated()
             logging.info(
                 "SAST report with id %s is still not created. "
-                "Next check in %s second(s). There are %s attempts left",
-                self._sast_report_id, wait_seconds, repeat_nr)
+                "Next check in %s second(s). There are %s attempts remaining before giving up.",
+                sast_report_id, wait_seconds, repeat_nr)
             time.sleep(wait_seconds)
 
     def datetime(self, *metric_source_ids: str) -> DateTime:
@@ -206,11 +218,12 @@ class Checkmarx(MetricSourceWithIssues):
         dates = []
         for project_name in metric_source_ids:
             try:
-                last_scan_json = self._fetch_last_scan(project_name)
+                project_id = self._fetch_project_id(project_name)
+                last_scan_json = self._fetch_last_scan(project_id)
                 dates.append(self.__parse_datetime(last_scan_json))
             except url_opener.UrlOpener.url_open_exceptions:
                 return datetime.datetime.min
-            except (KeyError, IndexError) as reason:
+            except (KeyError, IndexError, ValueError) as reason:
                 logging.error("Couldn't parse date and time for project %s from %s: %s",
                               project_name, self.url(), reason)
                 return datetime.datetime.min
@@ -220,28 +233,23 @@ class Checkmarx(MetricSourceWithIssues):
     def __parse_datetime(json_object: Dict) -> DateTime:
         """ Parse the JSON to get the date and time of the last scan. """
         datetimes = []
-        datetime_string = json_object["ScanCompletedOn"].split('.')[0]
+        datetime_string = json_object[0]["dateAndTime"]["finishedOn"].split('.')[0]
         datetimes.append(dateutil.parser.parse(datetime_string, ignoretz=True))
-        comment = json_object.get("Comment", "")
-        comment_sep, prefix, postfix = '; ', 'Attempt to perform scan on ', ' - No code changes were detected'
-        if comment_sep in comment:
-            for check in comment.strip(comment_sep).split(comment_sep):
-                if prefix in check:
-                    datetime_string = check.strip(prefix).strip(postfix)
-                    datetimes.append(dateutil.parser.parse(datetime_string, dayfirst=False))
         return max(datetimes, default=datetime.datetime.min)
 
     @functools.lru_cache(maxsize=1024)
-    def _fetch_last_scan(self, project_name: str) -> Dict[str, int]:
+    def _fetch_project_id(self, project_name: str) -> int:
+        projects = self.__get_json(self._projects_url)
+        return next(project['id'] for project in projects if project['name'] == project_name)
+
+    def _fetch_statistics(self, scan_id: int) -> Dict[str, int]:
+        return self.__get_json(self._statistics_url.format(id=scan_id))
+
+    @functools.lru_cache(maxsize=1024)
+    def _fetch_last_scan(self, project_d: int) -> Dict[str, int]:
         """ Create the api URL and fetch the report from it. """
-        api_url = self._last_scan_url.format(project_name=urllib.parse.quote(project_name))
-        return self.__get_json(api_url)["value"][0]["LastScan"]
+        return self.__get_json(self._scan_url.format(project_id=project_d))
 
-    def __get_json(self, api_url: str, opener: url_opener.UrlOpener = None) -> Dict:
+    def __get_json(self, api_url: str) -> Dict:
         """ Return and evaluate the JSON at the url using Basic Authentication. """
-
-        if opener is None:
-            opener = self._url_open
-        json_string = opener.url_read(api_url)
-
-        return utils.eval_json(json_string)
+        return utils.eval_json(self._url_open.url_read(api_url))